requirements.txt
# Core LLM and Agent Framework
langchain==0.3.13
langchain-community==0.3.13
langchain-ollama==0.2.2
langgraph==0.2.62
langsmith==0.2.7

# Vector Database and Embeddings
chromadb==0.5.23
sentence-transformers==3.3.1

# Document Processing
pymupdf==1.24.14
pypdf==5.1.0
python-docx==1.2.0
unstructured==0.16.15

# Memory Systems
redis==5.2.1
neo4j==5.27.0

# API and Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic==2.10.4
pydantic-settings==2.7.1

# UI
streamlit==1.41.1
plotly==5.24.1

# Evaluation
ragas==0.2.14
datasets==3.2.0

# Utilities
python-dotenv==1.0.1
loguru==0.7.3
tenacity==9.0.0
httpx==0.28.1

# Development and Testing
pytest==8.3.4
pytest-asyncio==0.24.0
black==24.10.0
flake8==7.1.1
mypy==1.14.1

# Monitoring
prometheus-client==0.21.1
